fix: do include quotes in filter strings, but only for the lerna commands
The top level npm-check-updates calls need their filter strings not to be double quoted, since it's a normal shell. When lerna executes the call for the lerna packages though it does need the strings to be double quoted as they're first parsed by lerna before being passed to the shell.